+++ forui/lib/widgets/slider.dart
@@ -1,7 +1,9 @@
 library forui.widgets.slider;
 
 export '../src/widgets/slider/old/slider.dart' hide InheritedData;
-export '../src/widgets/slider/old/slider_controller.dart' hide GestureAdjustment, KeyboardAdjustment;
-export '../src/widgets/slider/old/slider_mark.dart';
-export '../src/widgets/slider/slider_value.dart' hide RawValue;
-export '../src/widgets/slider/old/thumb.dart' hide Thumb;
+export '../src/widgets/slider/thumb.dart' hide Thumb;
+
+export '../src/widgets/slider/slider.dart';
+export '../src/widgets/slider/slider_controller.dart';
+export '../src/widgets/slider/slider_mark.dart';
+export '../src/widgets/slider/slider_selection.dart' show FSliderSelection;
